backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
]
